LPS-199599 Inline MiniumLayoutsInitializer methods into MiniumSiteInitializer
